clustermesh,service: Relocate store package to clustermesh
After the removal of external workloads the only user for ClusterService
is ClusterMesh. Move pkg/service/store to pkg/clustermesh/store to properly
reflect use and make it easier to clean up pkg/service.
